+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Changed
 
 - Updated argent class hash used in account creation to v0.4.0
+- Temporarily disabled integration with braavos accounts
 
 #### Removed